requirements.txt
python-dotenv==1.0.1
pydantic==2.9.2
marker-pdf==1.0.2
timm==1.0.11
einops==0.8.0
uvicorn==0.32.0
fastapi==0.115.4
python-multipart==0.0.17